backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Werkzeug==2.3.7
Pillow==10.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
